requirements.txt
praw==7.8.1
pandas
matplotlib==3.8.2
seaborn==0.13.0
tkinter-tooltip==2.1.0
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.33.0
streamlit-local-storage
